Refactor OAuth2 client auth and move /me endpoints to /v1/me
- Refactor OAuth2 token endpoint to require client credentials via HTTP
  Basic Auth, removing `client_id` and `client_secret` from the form
  body for improved RFC 6749 compliance
- Move `/me`, `/me/organizations`, and `/me/projects` endpoints from
  the root router to a new `/v1/me` subrouter for better API structure
- Update tests to use `/v1/me` endpoints and HTTP Basic Auth for token
  requests; split `/me` endpoint tests into a dedicated test file
- Remove `client_id` and `client_secret` from the `TokenRequest` model
- Add docstring to OAuth client credentials dependency for clarity
