+++ any_auth/api/v1/routes.py
@@ -2,6 +2,7 @@
 import fastapi
 
 from any_auth.api.v1.aks import router as api_keys_router
+from any_auth.api.v1.me.routes import router as me_router
 from any_auth.api.v1.oauth import router as oauth_router
 from any_auth.api.v1.org_mem_ras import router as org_mem_rs_router
 from any_auth.api.v1.org_mems import router as org_members_router
@@ -18,6 +19,7 @@ from any_auth.api.v1.users import router as users_router
 
 router = fastapi.APIRouter(prefix="/v1")
 
+router.include_router(me_router)
 router.include_router(api_keys_router)
 router.include_router(oauth_router)
 router.include_router(org_mem_rs_router)
